gst-msdk: move common code to lib
