+++ test/gst-msdk/decode/10bit/hevc.py
@@ -5,8 +5,8 @@
 ###
 
 from .....lib import *
-from ...util import *
-from ..decoder import DecoderTest
+from .....lib.gstreamer.msdk.util import *
+from .....lib.gstreamer.msdk.decoder import DecoderTest
 
 spec = load_test_spec("hevc", "decode", "10bit")
