+++ test/gst-msdk/decode/vc1.py
@@ -5,8 +5,8 @@
 ###
 
 from ....lib import *
-from ..util import *
-from .decoder import DecoderTest
+from ....lib.gstreamer.msdk.util import *
+from ....lib.gstreamer.msdk.decoder import DecoderTest
 
 spec = load_test_spec("vc1", "decode")
 spec_r2r = load_test_spec("vc1", "decode", "r2r")
