+++ test/gst-msdk/encode/mpeg2.py
@@ -5,8 +5,8 @@
 ###
 
 from ....lib import *
-from ..util import *
-from .encoder import EncoderTest
+from ....lib.gstreamer.msdk.util import *
+from ....lib.gstreamer.msdk.encoder import EncoderTest
 
 spec = load_test_spec("mpeg2", "encode")
 spec_r2r = load_test_spec("mpeg2", "encode", "r2r")
