+++ test/gst-msdk/transcode/vc1.py
@@ -5,8 +5,8 @@
 ###
 
 from ....lib import *
-from ..util import *
-from .transcoder import TranscoderTest
+from ....lib.gstreamer.msdk.util import *
+from ....lib.gstreamer.msdk.transcoder import TranscoderTest
 
 spec = load_test_spec("vc1", "transcode")
